apf_run: Fix ram_len adjustment for insufficient data_len
Previously, the logic was if data_len didn't accomodate the
hardcoded counters for the interpreter, add that size to the
ram_len.

This adds the hardcoded counter size regardless of how much
actually needs to be added. E.g. if 16 data bytes were provided
but 20 were required for the counters, this would result in 36
bytes, not 20.

Now, the logic is to add only the exact amount required so
data_len can accomodate the hardcoded counters for the
interpreter.

Additionally, this is now done before ram_len is aligned to a
4-byte boundary. Previously, with the adjustment occuring after
aligning ram_len, this could un-align ram_len if data_len was
itself not aligned (e.g. from command line argument).

Bug: 419135813
Test: m apf_run && atest apf_run_test
Change-Id: Ia4f7b0ede73a4ee9e97c1997c776e7ff7629419b